+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Added `BearerTokenChallengeAuthenticationPolicy`, which enables creation of authentication policies that can handle challenges.
+- Added additional methods to `BearerTokenAuthenticationPolicy`, which enables creation of authentication policies that can handle challenges.
 
 ## 1.13.0 (2021-04-07)
