+++ sdk/core/Azure.Core/src/Pipeline/BearerTokenAuthenticationPolicy.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.Diagnostics;
@@ -15,6 +16,7 @@ namespace Azure.Core.Pipeline
     /// </summary>
     public class BearerTokenAuthenticationPolicy : HttpPipelinePolicy
     {
+        private string[] _scopes;
         private readonly AccessTokenCache _accessTokenCache;
 
         /// <summary>
@@ -28,16 +30,21 @@ namespace Azure.Core.Pipeline
         /// Creates a new instance of <see cref="BearerTokenAuthenticationPolicy"/> using provided token credential and scopes to authenticate for.
         /// </summary>
         /// <param name="credential">The token credential to use for authentication.</param>
-        /// <param name="scopes">Scopes to authenticate for.</param>
+        /// <param name="scopes">_scopes to authenticate for.</param>
         public BearerTokenAuthenticationPolicy(TokenCredential credential, IEnumerable<string> scopes)
-            : this(credential, scopes, TimeSpan.FromMinutes(5), TimeSpan.FromSeconds(30)) { }
+            : this(credential, scopes, TimeSpan.FromMinutes(5), TimeSpan.FromSeconds(30))
+        { }
 
-        internal BearerTokenAuthenticationPolicy(TokenCredential credential, IEnumerable<string> scopes, TimeSpan tokenRefreshOffset, TimeSpan tokenRefreshRetryDelay)
+        internal BearerTokenAuthenticationPolicy(TokenCredential credential,
+            IEnumerable<string> scopes,
+            TimeSpan tokenRefreshOffset,
+            TimeSpan tokenRefreshRetryDelay)
         {
             Argument.AssertNotNull(credential, nameof(credential));
             Argument.AssertNotNull(scopes, nameof(scopes));
 
-            _accessTokenCache = new AccessTokenCache(credential, scopes.ToArray(), tokenRefreshOffset, tokenRefreshRetryDelay);
+            _scopes = scopes.ToArray();
+            _accessTokenCache = new AccessTokenCache(credential, tokenRefreshOffset, tokenRefreshRetryDelay, scopes.ToArray());
         }
 
         /// <inheritdoc />
@@ -52,6 +59,55 @@ namespace Azure.Core.Pipeline
             ProcessAsync(message, pipeline, false).EnsureCompleted();
         }
 
+        /// <summary>
+        /// Executes before <see cref="ProcessAsync(HttpMessage, ReadOnlyMemory{HttpPipelinePolicy})"/> or
+        /// <see cref="Process(HttpMessage, ReadOnlyMemory{HttpPipelinePolicy})"/> is called.
+        /// Implementers of this method are expected to call <see cref="SetAuthorizationHeader"/> or <see cref="SetAuthorizationHeaderAsync"/>
+        /// if authorization is required for requests not related to handling a challenge response.
+        /// </summary>
+        /// <param name="message">The <see cref="HttpMessage"/> this policy would be applied to.</param>
+        /// <returns>The <see cref="ValueTask"/> representing the asynchronous operation.</returns>
+        protected virtual ValueTask AuthorizeRequestAsync(HttpMessage message)
+        {
+            var context = new TokenRequestContext(_scopes, message.Request.ClientRequestId);
+            return SetAuthorizationHeaderAsync(message, context);
+        }
+
+        /// <summary>
+        /// Executes before <see cref="ProcessAsync(HttpMessage, ReadOnlyMemory{HttpPipelinePolicy})"/> or
+        /// <see cref="Process(HttpMessage, ReadOnlyMemory{HttpPipelinePolicy})"/> is called.
+        /// Implementers of this method are expected to call <see cref="SetAuthorizationHeader"/> or <see cref="SetAuthorizationHeaderAsync"/>
+        /// if authorization is required for requests not related to handling a challenge response.
+        /// </summary>
+        /// <param name="message">The <see cref="HttpMessage"/> this policy would be applied to.</param>
+        protected virtual void AuthorizeRequest(HttpMessage message)
+        {
+            var context = new TokenRequestContext(_scopes, message.Request.ClientRequestId);
+            SetAuthorizationHeader(message, context);
+        }
+
+        /// <summary>
+        /// Executed in the event a 401 response with a WWW-Authenticate authentication challenge header is received after the initial request.
+        /// </summary>
+        /// <remarks>Service client libraries may override this to handle service specific authentication challenges.</remarks>
+        /// <param name="message">The <see cref="HttpMessage"/> to be authenticated.</param>
+        /// <returns>A boolean indicating whether the request was successfully authenticated and should be sent to the transport.</returns>
+        protected virtual ValueTask<bool> AuthorizeRequestOnChallengeAsync(HttpMessage message)
+        {
+            return default;
+        }
+
+        /// <summary>
+        /// Executed in the event a 401 response with a WWW-Authenticate authentication challenge header is received after the initial request.
+        /// </summary>
+        /// <remarks>Service client libraries may override this to handle service specific authentication challenges.</remarks>
+        /// <param name="message">The <see cref="HttpMessage"/> to be authenticated.</param>
+        /// <returns>A boolean indicating whether the request was successfully authenticated and should be sent to the transport.</returns>
+        protected virtual bool AuthorizeRequestOnChallenge(HttpMessage message)
+        {
+            return false;
+        }
+
         private async ValueTask ProcessAsync(HttpMessage message, ReadOnlyMemory<HttpPipelinePolicy> pipeline, bool async)
         {
             if (message.Request.Uri.Scheme != Uri.UriSchemeHttps)
@@ -59,38 +115,82 @@ namespace Azure.Core.Pipeline
                 throw new InvalidOperationException("Bearer token authentication is not permitted for non TLS protected (https) endpoints.");
             }
 
-            string headerValue = await _accessTokenCache.GetHeaderValueAsync(message, async).ConfigureAwait(false);
-            message.Request.SetHeader(HttpHeader.Names.Authorization, headerValue);
-
             if (async)
             {
+                await AuthorizeRequestAsync(message).ConfigureAwait(false);
                 await ProcessNextAsync(message, pipeline).ConfigureAwait(false);
             }
             else
             {
+                AuthorizeRequest(message);
                 ProcessNext(message, pipeline);
             }
+
+            // Check if we have received a challenge or we have not yet issued the first request.
+            if (message.Response.Status == (int)HttpStatusCode.Unauthorized && message.Response.Headers.Contains(HttpHeader.Names.WwwAuthenticate))
+            {
+                // Attempt to get the TokenRequestContext based on the challenge.
+                // If we fail to get the context, the challenge was not present or invalid.
+                // If we succeed in getting the context, authenticate the request and pass it up the policy chain.
+                if (async)
+                {
+                    if (await AuthorizeRequestOnChallengeAsync((message)).ConfigureAwait(false))
+                    {
+                        await ProcessNextAsync(message, pipeline).ConfigureAwait(false);
+                    }
+                }
+                else
+                {
+                    if (AuthorizeRequestOnChallenge(message))
+                    {
+                        ProcessNext(message, pipeline);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Sets the Authorization header on the <see cref="Request"/>.
+        /// </summary>
+        /// <param name="message">The <see cref="HttpMessage"/> with the <see cref="Request"/> to be authorized.</param>
+        /// <param name="context">The <see cref="TokenRequestContext"/> used to authorize the <see cref="Request"/>.</param>
+        protected async ValueTask SetAuthorizationHeaderAsync(HttpMessage message, TokenRequestContext context)
+        {
+            string headerValue = await _accessTokenCache.GetHeaderValueAsync(message, context, true).ConfigureAwait(false);
+            message.Request.Headers.SetValue(HttpHeader.Names.Authorization, headerValue);
+        }
+
+        /// <summary>
+        /// Sets the Authorization header on the <see cref="Request"/>.
+        /// </summary>
+        /// <param name="message">The <see cref="HttpMessage"/> with the <see cref="Request"/> to be authorized.</param>
+        /// <param name="context">The <see cref="TokenRequestContext"/> used to authorize the <see cref="Request"/>.</param>
+        protected void SetAuthorizationHeader(HttpMessage message, TokenRequestContext context)
+        {
+            string headerValue = _accessTokenCache.GetHeaderValueAsync(message, context, false).EnsureCompleted();
+            message.Request.Headers.SetValue(HttpHeader.Names.Authorization, headerValue);
         }
 
         private class AccessTokenCache
         {
             private readonly object _syncObj = new object();
             private readonly TokenCredential _credential;
-            private readonly string[] _scopes;
             private readonly TimeSpan _tokenRefreshOffset;
             private readonly TimeSpan _tokenRefreshRetryDelay;
 
+            private TokenRequestContext? _currentContext;
             private TaskCompletionSource<HeaderValueInfo>? _infoTcs;
             private TaskCompletionSource<HeaderValueInfo>? _backgroundUpdateTcs;
-            public AccessTokenCache(TokenCredential credential, string[] scopes, TimeSpan tokenRefreshOffset, TimeSpan tokenRefreshRetryDelay)
+
+            public AccessTokenCache(TokenCredential credential, TimeSpan tokenRefreshOffset, TimeSpan tokenRefreshRetryDelay, string[] initialScopes)
             {
                 _credential = credential;
-                _scopes = scopes;
                 _tokenRefreshOffset = tokenRefreshOffset;
                 _tokenRefreshRetryDelay = tokenRefreshRetryDelay;
+                _currentContext = new TokenRequestContext(initialScopes);
             }
 
-            public async ValueTask<string> GetHeaderValueAsync(HttpMessage message, bool async)
+            public async ValueTask<string> GetHeaderValueAsync(HttpMessage message, TokenRequestContext context, bool async)
             {
                 bool getTokenFromCredential;
                 TaskCompletionSource<HeaderValueInfo> headerValueTcs;
@@ -99,30 +199,36 @@ namespace Azure.Core.Pipeline
 
                 while (true)
                 {
-                    (headerValueTcs, backgroundUpdateTcs, getTokenFromCredential) = GetTaskCompletionSources();
+                    (headerValueTcs, backgroundUpdateTcs, getTokenFromCredential) = GetTaskCompletionSources(context);
+                    HeaderValueInfo info;
                     if (getTokenFromCredential)
                     {
                         if (backgroundUpdateTcs != null)
                         {
-#pragma warning disable AZC0111 // DO NOT use EnsureCompleted in possibly asynchronous scope.
-                            HeaderValueInfo info = headerValueTcs.Task.EnsureCompleted();
-#pragma warning restore AZC0111 // DO NOT use EnsureCompleted in possibly asynchronous scope.
-                            _ = Task.Run(() => GetHeaderValueFromCredentialInBackgroundAsync(backgroundUpdateTcs, info, message, async));
+                            if (async)
+                            {
+                                info = await headerValueTcs.Task.ConfigureAwait(false);
+                            }
+                            else
+                            {
+#pragma warning disable AZC0104 // Use EnsureCompleted() directly on asynchronous method return value.
+                                info = headerValueTcs.Task.EnsureCompleted();
+#pragma warning restore AZC0104 // Use EnsureCompleted() directly on asynchronous method return value.
+                            }
+                            _ = Task.Run(() => GetHeaderValueFromCredentialInBackgroundAsync(backgroundUpdateTcs, info, context, async));
                             return info.HeaderValue;
                         }
 
                         try
                         {
-                            HeaderValueInfo info = await GetHeaderValueFromCredentialAsync(message, async, message.CancellationToken).ConfigureAwait(false);
+                            info = await GetHeaderValueFromCredentialAsync(context, async, message.CancellationToken).ConfigureAwait(false);
                             headerValueTcs.SetResult(info);
                         }
                         catch (OperationCanceledException)
                         {
                             headerValueTcs.SetCanceled();
                         }
-#pragma warning disable CA1031 // Catch more specific allowed exception type, or rethrow the exception
                         catch (Exception exception)
-#pragma warning restore CA1031 // Catch more specific allowed exception type, or rethrow the exception
                         {
                             headerValueTcs.SetException(exception);
                             // The exception will be thrown on the next lines when we touch the result of
@@ -148,12 +254,19 @@ namespace Azure.Core.Pipeline
                                 catch (AggregateException) { } // ignore exception here to rethrow it with EnsureCompleted
                             }
                         }
+                        if (async)
+                        {
+                            info = await headerValueTcs.Task.ConfigureAwait(false);
+                        }
+                        else
+                        {
+#pragma warning disable AZC0104 // Use EnsureCompleted() directly on asynchronous method return value.
+                            info = headerValueTcs.Task.EnsureCompleted();
+#pragma warning restore AZC0104 // Use EnsureCompleted() directly on asynchronous method return value.
+                        }
 
-#pragma warning disable AZC0111 // DO NOT use EnsureCompleted in possibly asynchronous scope.
-                        return headerValueTcs.Task.EnsureCompleted().HeaderValue;
-#pragma warning restore AZC0111 // DO NOT use EnsureCompleted in possibly asynchronous scope.
+                        return info.HeaderValue;
                     }
-
                     catch (TaskCanceledException) when (!message.CancellationToken.IsCancellationRequested)
                     {
                         maxCancellationRetries--;
@@ -171,15 +284,18 @@ namespace Azure.Core.Pipeline
                 }
             }
 
-            private (TaskCompletionSource<HeaderValueInfo> Tcs, TaskCompletionSource<HeaderValueInfo>? BackgroundUpdateTcs, bool GetTokenFromCredential) GetTaskCompletionSources()
+            private (TaskCompletionSource<HeaderValueInfo> InfoTcs, TaskCompletionSource<HeaderValueInfo>? BackgroundUpdateTcs, bool GetTokenFromCredential)
+                GetTaskCompletionSources(TokenRequestContext context)
             {
                 lock (_syncObj)
                 {
                     // Initial state. GetTaskCompletionSources has been called for the first time
-                    if (_infoTcs == null)
+                    if (_infoTcs == null || RequestRequiresNewToken(context))
                     {
+                        _currentContext = context;
                         _infoTcs = new TaskCompletionSource<HeaderValueInfo>(TaskCreationOptions.RunContinuationsAsynchronously);
-                        return (_infoTcs, default, true);
+                        _backgroundUpdateTcs = default;
+                        return (_infoTcs, _backgroundUpdateTcs, true);
                     }
 
                     // Getting new access token is in progress, wait for it
@@ -191,7 +307,9 @@ namespace Azure.Core.Pipeline
 
                     DateTimeOffset now = DateTimeOffset.UtcNow;
                     // Access token has been successfully acquired in background and it is not expired yet, use it instead of current one
-                    if (_backgroundUpdateTcs != null && _backgroundUpdateTcs.Task.Status == TaskStatus.RanToCompletion && _backgroundUpdateTcs.Task.Result.ExpiresOn > now)
+                    if (_backgroundUpdateTcs != null &&
+                        _backgroundUpdateTcs.Task.Status == TaskStatus.RanToCompletion &&
+                        _backgroundUpdateTcs.Task.Result.ExpiresOn > now)
                     {
                         _infoTcs = _backgroundUpdateTcs;
                         _backgroundUpdateTcs = default;
@@ -216,23 +334,32 @@ namespace Azure.Core.Pipeline
                 }
             }
 
-            private async ValueTask GetHeaderValueFromCredentialInBackgroundAsync(TaskCompletionSource<HeaderValueInfo> backgroundUpdateTcs, HeaderValueInfo info, HttpMessage httpMessage, bool async)
+            // must be called under lock (_syncObj)
+            private bool RequestRequiresNewToken(TokenRequestContext context) =>
+                _currentContext == null ||
+                (context.Scopes != null && !context.Scopes.AsSpan().SequenceEqual(_currentContext.Value.Scopes.AsSpan())) ||
+                (context.Claims != null && !string.Equals(context.Claims, _currentContext.Value.Claims));
+
+            private async ValueTask GetHeaderValueFromCredentialInBackgroundAsync(TaskCompletionSource<HeaderValueInfo> backgroundUpdateTcs,
+                HeaderValueInfo info,
+                TokenRequestContext context,
+                bool async)
             {
                 var cts = new CancellationTokenSource(_tokenRefreshRetryDelay);
                 try
                 {
-                    HeaderValueInfo newInfo = await GetHeaderValueFromCredentialAsync(httpMessage, async, cts.Token).ConfigureAwait(false);
+                    HeaderValueInfo newInfo = await GetHeaderValueFromCredentialAsync(context, async, cts.Token).ConfigureAwait(false);
                     backgroundUpdateTcs.SetResult(newInfo);
                 }
                 catch (OperationCanceledException oce) when (cts.IsCancellationRequested)
                 {
                     backgroundUpdateTcs.SetResult(new HeaderValueInfo(info.HeaderValue, info.ExpiresOn, DateTimeOffset.UtcNow));
-                    AzureCoreEventSource.Singleton.BackgroundRefreshFailed(httpMessage.Request.ClientRequestId, oce.ToString());
+                    AzureCoreEventSource.Singleton.BackgroundRefreshFailed(context.ParentRequestId ?? string.Empty, oce.ToString());
                 }
                 catch (Exception e)
                 {
                     backgroundUpdateTcs.SetResult(new HeaderValueInfo(info.HeaderValue, info.ExpiresOn, DateTimeOffset.UtcNow + _tokenRefreshRetryDelay));
-                    AzureCoreEventSource.Singleton.BackgroundRefreshFailed(httpMessage.Request.ClientRequestId, e.ToString());
+                    AzureCoreEventSource.Singleton.BackgroundRefreshFailed(context.ParentRequestId ?? string.Empty, e.ToString());
                 }
                 finally
                 {
@@ -240,12 +367,11 @@ namespace Azure.Core.Pipeline
                 }
             }
 
-            private async ValueTask<HeaderValueInfo> GetHeaderValueFromCredentialAsync(HttpMessage message, bool async, CancellationToken cancellationToken)
+            private async ValueTask<HeaderValueInfo> GetHeaderValueFromCredentialAsync(TokenRequestContext context, bool async, CancellationToken cancellationToken)
             {
-                var requestContext = new TokenRequestContext(_scopes, message.Request.ClientRequestId);
                 AccessToken token = async
-                    ? await _credential.GetTokenAsync(requestContext, cancellationToken).ConfigureAwait(false)
-                    : _credential.GetToken(requestContext, cancellationToken);
+                    ? await _credential.GetTokenAsync(context, cancellationToken).ConfigureAwait(false)
+                    : _credential.GetToken(context, cancellationToken);
 
                 return new HeaderValueInfo("Bearer " + token.Token, token.ExpiresOn, token.ExpiresOn - _tokenRefreshOffset);
             }
